Create Graph Detection Algorithm
How to detect if two graphs or pictures are the same using C++. The algorithm checks the differences between the two graphs/pictures and returns the amount of differences. The user can then choose to ignore a certain amount of differences between the two algorithms by entering a number of allowances.