Sudoku Solving Algorithm
An algorithm that solves sudoku problems. Sudoku problem must be manually entered into the code where 0s are empty spots and the numbers are actual numbers in the sudoku puzzle. Algorithm could be easily tweaked to allow the user to input the numbers of the sudoku puzzle through the console, but that would take a fairly long time.